get_info_doutor.pyを修正
